requirements.txt
Flask==3.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0
qrcode==7.4.2
Pillow==10.1.0
werkzeug==3.0.1